convert numrows from int to int64
most of the code has already moved to int64, but a couple interfaces were casting to int intermittently

+++ flow/connectors/clickhouse/cdc.go
@@ -100,7 +100,7 @@ func (c *ClickHouseConnector) syncRecordsViaAvro(
 
 	return &model.SyncResponse{
 		LastSyncedCheckpoint: req.Records.GetLastCheckpoint(),
-		NumRecordsSynced:     int64(numRecords),
+		NumRecordsSynced:     numRecords,
 		CurrentSyncBatchID:   syncBatchID,
 		TableNameRowsMapping: tableNameRowsMapping,
 		TableSchemaDeltas:    req.Records.SchemaDeltas,
