+++ flow/connectors/s3/s3.go
@@ -113,7 +113,7 @@ func (c *S3Connector) SyncRecords(ctx context.Context, req *model.SyncRecordsReq
 
 	return &model.SyncResponse{
 		LastSyncedCheckpoint: lastCheckpoint,
-		NumRecordsSynced:     int64(numRecords),
+		NumRecordsSynced:     numRecords,
 		TableNameRowsMapping: tableNameRowsMapping,
 		TableSchemaDeltas:    req.Records.SchemaDeltas,
 	}, nil
